Mu2eII_SM21: Merge head of Offline including new physics generators

+++ RecoDataProducts/inc/CrvDigi.hh
@@ -19,22 +19,22 @@ namespace mu2e
 
     CrvDigi() {}
 
-    CrvDigi(const std::array<unsigned int, NSamples> &ADCs, unsigned int startTDC, mu2e::CRSScintillatorBarIndex scintillatorBarIndex, int SiPMNumber) :
+    CrvDigi(const std::array<int16_t, NSamples> &ADCs, uint16_t startTDC, mu2e::CRSScintillatorBarIndex scintillatorBarIndex, uint8_t SiPMNumber) :
             _ADCs(ADCs), _startTDC(startTDC), _scintillatorBarIndex(scintillatorBarIndex), _SiPMNumber(SiPMNumber) {}
 
-    const std::array<unsigned int, NSamples> &GetADCs() const     {return _ADCs;}
-    unsigned int                              GetStartTDC() const {return _startTDC;}
+    const std::array<int16_t, NSamples>  &GetADCs() const     {return _ADCs;}
+    uint16_t                              GetStartTDC() const {return _startTDC;}
 
     mu2e::CRSScintillatorBarIndex GetScintillatorBarIndex() const {return _scintillatorBarIndex;}
-    int                           GetSiPMNumber() const           {return _SiPMNumber;}
+    uint8_t                       GetSiPMNumber() const           {return _SiPMNumber;}
 
     private:
 
-    std::array<unsigned int, NSamples> _ADCs;
-    unsigned int                       _startTDC;
+    std::array<int16_t, NSamples>  _ADCs;
+    uint16_t                       _startTDC;
 
     mu2e::CRSScintillatorBarIndex  _scintillatorBarIndex;
-    int                            _SiPMNumber; 
+    uint8_t                        _SiPMNumber; 
   };
   typedef std::vector<mu2e::CrvDigi> CrvDigiCollection;
 }
